my-project/requirements.txt
alpha-vantage==2.3.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
gspread==5.10.0
oauth2client==4.1.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
python-telegram-bot==20.4
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
schedule==1.2.0
loguru==0.7.0
plotly==5.15.0
streamlit==1.25.0
psutil
gspread
loguru
